% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/construction.R
\docType{data}
\name{construction}
\alias{construction}
\title{Construction}
\format{Data frame with columns
\describe{
\item{period}{Date Monthly}
\item{homesManufactured}{numeric Thousands of units}
\item{housingPermits}{numeric Thousands of units}
\item{housingStarts}{numeric Thousands of units}
\item{newHomeSales}{numeric Thousands of units}
}}
\source{
\url{http://www.census.gov/econ/currentdata/}
}
\usage{
construction
}
\description{
New home permits, starts, and sales
}
\examples{
data(construction)
}
\keyword{datasets}

