% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Errors.R
\name{CovMatrix}
\alias{CovMatrix}
\title{Find the covariance matrix for an input linear regression model}
\usage{
CovMatrix(Model, errors = c("homoskedastic", "robust", "white", "HC0", "HC2",
  "HC3", "HC4", "HC4m", "HC5"), cluster = FALSE)
}
\arguments{
\item{Model}{A linear model from the \code{\link{lm}} command.}

\item{errors}{The type of standard errors. The options are c("homoskedastic", "robust", "white", "HC0" , "HC2" , "HC3" , "HC4", "HC4m", "HC5") where the last ones are as described by \code{\link{vcovHC}}}

\item{cluster}{The clustering variable. If input errors (above) are ignored.}
}
\value{
The covariance matrix for the input linear model with the standard errors as specified by the errors parameter.
}
\description{
Find the covariance matrix for an input linear regression model
}
\examples{
x = sin(1:100)
y = 1 + x + c(rnorm(50), rnorm(50, sd = 3))
z = as.factor(c(rep("A", 50), rep("B", 50) ))
dd = data.frame(x = sin(1:100), y = y, z = z)
form = "y ~ x"
rm(x,y,z)
M1  = lm(form = form, data = dd)
Cov = CovMatrix(M1, cluster = dd$z)
}
\seealso{
\code{\link{lm}}; \code{\link{vcovHC}}
}

