% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralFunctions.R
\name{gazer}
\alias{gazer}
\title{This is a wrapper function for the stargazer package so that econreg objects can easily be passed to it.}
\usage{
gazer(models, type = "text", digits = 7, ...)
}
\arguments{
\item{models}{A econreg object or a list of econreg objects.}

\item{...}{Any other options for the stargazer package. See  \code{\link{stargazer}} for descriptions of these.}
}
\value{
This does not return anything. The table is either printed to the console or saved on the filing system depending on what options are input.
}
\description{
This is a wrapper function for the stargazer package so that econreg objects can easily be passed to it.
}
\examples{
x = sin(1:100)
y = 1 + x + c(rnorm(50), rnorm(50, sd = 3))
z = as.factor(c(rep("A", 50), rep("B", 50) ))
dd = data.frame(x = sin(1:100), y = y, z = z)
form = "y ~ x"
rm(x,y,z)
M1 = reg(form = form, data = dd, cluster = "z")
gazer(M1)
M2 = reg(form = form, data = dd, errors = "homoskedastic")
models = list(M1, M2)
gazer(models)
}
\seealso{
\code{\link{stargazer}}
}

