% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/housing.R
\docType{data}
\name{housing}
\alias{housing}
\title{Housing}
\format{Data frame with columns
\describe{
\item{period}{Date Quarterly}
\item{homeownership}{numeric Percentage}
\item{ownerVacancy}{numeric Percentage}
\item{rentalVacancy}{numeric Percentage}
}}
\source{
\url{http://www.census.gov/econ/currentdata/}
}
\usage{
housing
}
\description{
Ownership and vacancy rates of US homes
}
\examples{
data(housing)
}
\keyword{datasets}

