% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nsror.R
\docType{data}
\name{nsror}
\alias{nsror}
\title{Net sales, receipts, and operating revenues}
\format{Data frame with columns
\describe{
\item{period}{Date Quarterly}
\item{nsror}{numeric total millions USD}
\item{manufacturing}{numeric Millions USD}
\item{mining}{numeric Millions USD}
\item{profServices}{numeric Millions USD}
\item{retail}{numeric Millions USD}
}}
\source{
\url{http://www.census.gov/econ/currentdata/}
}
\usage{
nsror
}
\description{
Select US industries
}
\examples{
data(nsror)
}
\keyword{datasets}

