% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/salesinventories.R
\docType{data}
\name{salesinventories}
\alias{salesinventories}
\title{Sales and inventories}
\format{Data frame with columns
\describe{
\item{period}{Date Monthly}
\item{balancePayments}{numeric Balance of payments in international trade}
\item{durableGoodsSales}{numeric Millions USD}
\item{durableGoodsInventories}{numeric Millions USD}
\item{manufacturersInventories}{numeric Millions USD}
\item{manufacturersNewOrders}{numeric Millions USD}
\item{manufacturersShipments}{numeric Millions USD}
\item{nondurableGoodsInventories}{numeric Millions USD}
\item{nondurableGoodsSales}{numeric Millions USD}
\item{retailFoodAdvanceSales}{numeric Millions USD}
\item{retailFoodSales}{numeric Millions USD}
\item{totalBusinessInventories}{numeric Millions USD}
\item{totalBusinessSales}{numeric Millions USD}
}}
\source{
\url{http://www.census.gov/econ/currentdata/}
}
\usage{
salesinventories
}
\description{
Sales and inventories of US economic sectors
}
\examples{
data(salesinventories)
}
\keyword{datasets}

