% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralFunctions.R
\name{SignificanceTests}
\alias{SignificanceTests}
\title{This is a function to estimate the tests of model significance with nonstandard standard errors.}
\usage{
SignificanceTests(Model, CovarianceMatrix, Class, digits = 4)
}
\arguments{
\item{Model}{The model.}

\item{CovarianceMatrix}{The Covariance Matrix}

\item{Class}{The class of model. Could be "lm", "probit".}

\item{digits}{How many digits to output the f statistic.}
}
\value{
Rreturns an F Statistic with stars for significance.
}
\description{
This is a function to estimate the tests of model significance with nonstandard standard errors.
}
\examples{
x = sin(1:100)
y = 1 + x + c(rnorm(50), rnorm(50, sd = 3))
z = as.factor(c(rep("A", 50), rep("B", 50) ))
dd = data.frame(x = sin(1:100), y = y, z = z)
form = "y ~ x"
rm(x,y,z)
M1 = reg(form = form, data = dd, cluster = "z")
SignificanceTests(M1$Model, M1$CovarianceMatrix, M1$Class, digits = 4)
}
\seealso{
\code{\link{stargazer}}
}

