% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinaryChoiceModels.R
\name{summary.econprobit}
\alias{summary.econprobit}
\title{This gives regression output for the econprobit model. Anova is output along with a coefficient test (with the specified errors), $R^2$, adjusted $R^2$ and residual standard error. Note that the residual standard error is done assuming homoskedastic standard errors. This is called by the \code{\link{reg}} function if \code{quietly = FALSE} or can be called manually.}
\usage{
\method{summary}{econprobit}(econreg)
}
\arguments{
\item{econprobit}{An econprobit object as created by the \code{\link{probit}} function.}
}
\value{
This does not return anything. A summary is printed to the console.
}
\description{
This gives regression output for the econprobit model. Anova is output along with a coefficient test (with the specified errors), $R^2$, adjusted $R^2$ and residual standard error. Note that the residual standard error is done assuming homoskedastic standard errors. This is called by the \code{\link{reg}} function if \code{quietly = FALSE} or can be called manually.
}
\examples{

}
\seealso{
\code{\link{anova}} \code{\link{summary.glm}} \code{\link{coeftest}}
}

