% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelRegression.R
\name{xtreg}
\alias{xtreg}
\title{Panel DataLinear Regression}
\usage{
xtreg(form, data, errors = c("homoskedastic", "robust", "white", "HC0", "HC2",
  "HC3", "HC4", "HC4m", "HC5"), cluster = FALSE, quietly = FALSE)
}
\arguments{
\item{form}{The regression formula. See \code{\link{xlm}} in the plm package for details on writing this formula.}

\item{data}{The data.frame for the regression. As subsetting is not currently supported you should drop observations you do not want to include from the data.frame.}

\item{errors}{The type of standard errors. The options are c("homoskedastic", "robust", "white", "HC0" , "HC2" , "HC3" , "HC4", "HC4m", "HC5") where the last ones are as described by \code{\link{sandwich::vcovHC}}}

\item{cluster}{The clustering variable. If input errors (above) are ignored.}

\item{quietly}{Output a summary? FALSE by default.}
}
\value{
An econreg object which contains a list:
 \enumerate{
 \item{"a linear model"}
 \item{"the covariance matrix"}
 \item{"the type of errors"}
 \item{"Class - The class of the model, lm"}
 }
}
\description{
If you get a names attribute error this can be caused by trying to do a within transformation on data where there is no withingroup variation in y.
}
\examples{
x = sin(1:100)
y = 1 + x + c(rnorm(50), rnorm(50, sd = 3))
z = as.factor(c(rep("A", 50), rep("B", 50) ))
dd = data.frame(x = x, y = y, z = z)
form = "y ~ x"
rm(x,y,z)
M1 = reg(form = form, data = dd, cluster = "z")
M2 = reg(form = form, data = dd, errors = "homoskedastic")
M3 = reg(form = form, data = dd, errors = "robust")
}
\seealso{
\code{\link{lm}}; \code{\link{sandwich::vcovHC}}
}

